deleted key vault interface
