+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/Secrets.java
@@ -3,5 +3,5 @@ package gov.hhs.cdc.trustedintermediary.wrappers;
 /** This interface is a blueprint for getting environment secrets */
 public interface Secrets {
 
-    String getKey();
+    String getKey(String secretName);
 }
